service/api/requirements.txt
absl-py==2.1.0
albucore==0.0.13
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
Automat==20.2.0
babel==2.16.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blessed==1.20.0
blinker==1.4
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
cloud-init==23.1.2
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
Cython==3.0.11
dbus-python==1.2.18
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.7.0
distro-info===1.1build1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
fastjsonschema==2.20.0
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.10.0
ftfy==6.3.1
gpustat==1.1.1
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
imageio==2.36.0
imgaug==0.4.0
importlib-metadata==4.6.4
incremental==21.3.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
json5==0.9.25
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
Levenshtein==0.26.1
lmdb==1.5.1
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
ninja==1.11.1.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
open_clip_torch==2.29.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.1
pyclipper==1.3.0.post6
pycparser==2.22
pycrdt==0.10.6
pycrdt-websocket==0.15.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
python-docx==1.1.2
python-json-logger==2.0.7
python-magic==0.4.24
python-multipart==0.0.17
pytz==2022.1
PyYAML==5.4.1
pyzmq==26.2.0
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.1
Send2Trash==1.8.3
service-identity==18.1.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sos==4.4
soupsieve==2.6
sqlite-anyio==0.2.3
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.1
systemd-python==234
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.11
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.0.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
triton==3.1.0
Twisted==22.1.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
uvicorn==0.32.0
virtualenv==20.27.1
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
zipp==1.0.0
zope.interface==5.4.0
